# api/proto/web/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: "3"
paths:
...
    /api/v1/emoticon/customize/create:
        post:
            tags:
                - Emoticon
            description: 创建表情包接口
            operationId: Emoticon_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EmoticonCreateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EmoticonCreateResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/emoticon/customize/delete:
        post:
            tags:
                - Emoticon
            description: 删除表情包接口
            operationId: Emoticon_Delete
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EmoticonDeleteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EmoticonDeleteResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/emoticon/customize/list:
        post:
            tags:
                - Emoticon
            description: 用户表情包列表接口
            operationId: Emoticon_List
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EmoticonListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EmoticonListResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /api/v1/message/delete:
        post:
            tags:
                - Message
            operationId: Message_Delete
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MessageDeleteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MessageDeleteResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/message/forward-records:
        post:
            tags:
                - Message
            description: 转发消息记录
            operationId: Message_ForwardRecords
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MessageForwardRecordsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MessageRecordsClearResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/message/history-records:
        post:
            tags:
                - Message
            description: 获取会话历史消息记录
            operationId: Message_HistoryRecords
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MessageHistoryRecordsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MessageHistoryRecordsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/message/records:
        post:
            tags:
                - Message
            description: 获取会话消息记录
            operationId: Message_Records
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MessageRecordsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MessageRecordsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/message/revoke:
        post:
            tags:
                - Message
            operationId: Message_Revoke
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MessageRevokeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MessageRevokeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        EmoticonCreateRequest:
            type: object
            properties:
                url:
                    type: string
            description: 创建表情包接口请求参数
        EmoticonCreateResponse:
            type: object
            properties:
                emoticonId:
                    type: integer
                    format: int32
                url:
                    type: string
            description: 创建表情包接口响应参数
        EmoticonDeleteRequest:
            type: object
            properties:
                emoticonId:
                    type: integer
                    format: int32
            description: 删除表情包接口请求参数
        EmoticonDeleteResponse:
            type: object
            properties: {}
            description: 删除表情包接口响应参数
        EmoticonItem:
            type: object
            properties:
                emoticonId:
                    type: integer
                    format: int32
                url:
                    type: string
        EmoticonListRequest:
            type: object
            properties: {}
            description: 用户表情包列表接口请求参数
        EmoticonListResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/EmoticonItem'
            description: 用户表情包列表接口响应参数
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        MessageDeleteRequest:
            required:
                - talkMode
                - toFromId
                - msgIds
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                toFromId:
                    type: integer
                    format: int32
                msgIds:
                    type: array
                    items:
                        type: string
        MessageDeleteResponse:
            type: object
            properties: {}
        MessageForwardRecordsRequest:
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                msgIds:
                    type: array
                    items:
                        type: string
        MessageHistoryRecordsRequest:
            required:
                - talkMode
                - toFromId
                - cursor
                - limit
                - msgType
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                toFromId:
                    type: integer
                    format: int32
                cursor:
                    type: integer
                    format: int32
                limit:
                    type: integer
                    format: int32
                msgType:
                    type: integer
                    format: int32
        MessageHistoryRecordsResponse:
            required:
                - items
                - cursor
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/MessageRecord'
                cursor:
                    type: integer
                    format: int32
        MessageRecord:
            required:
                - msgId
                - sequence
                - msgType
                - fromId
                - nickname
                - avatar
                - isRevoked
                - sendTime
                - extra
                - quote
            type: object
            properties:
                msgId:
                    type: string
                sequence:
                    type: integer
                    format: int32
                msgType:
                    type: integer
                    format: int32
                fromId:
                    type: integer
                    format: int32
                nickname:
                    type: string
                avatar:
                    type: string
                isRevoked:
                    type: integer
                    format: int32
                sendTime:
                    type: string
                extra:
                    type: string
                quote:
                    type: string
        MessageRecordsClearResponse:
            required:
                - items
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/MessageRecord'
        MessageRecordsRequest:
            required:
                - talkMode
                - toFromId
                - cursor
                - limit
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                toFromId:
                    type: integer
                    format: int32
                cursor:
                    type: integer
                    format: int32
                limit:
                    type: integer
                    format: int32
        MessageRecordsResponse:
            required:
                - items
                - cursor
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/MessageRecord'
                cursor:
                    type: integer
                    format: int32
        MessageRevokeRequest:
            required:
                - talkMode
                - toFromId
                - msgId
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                toFromId:
                    type: integer
                    format: int32
                msgId:
                    type: string
        MessageRevokeResponse:
            type: object
            properties: {}
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...